# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/product_prices:
    get:
      tags:
        - ProductPrices
      summary: List ProductPrices
      description: Get a paginated list of ProductPrices with optional filtering and sorting
      operationId: listProductPrices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPriceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ProductPrices
      summary: Create ProductPrice
      description: Create a new ProductPrice
      operationId: createProductPrice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductPriceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPrice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/recovery_invites:
    get:
      tags:
        - RecoveryInvites
      summary: List RecoveryInvites
      description: Get a paginated list of RecoveryInvites with optional filtering and sorting
      operationId: listRecoveryInvites
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryInviteList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RecoveryInvites
      summary: Create RecoveryInvite
      description: Create a new RecoveryInvite
      operationId: createRecoveryInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecoveryInviteRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryInvite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ProductPrice:
      type: object
      required:
        - id
        - website_id
        - item_id
        - list_price
        - currency
        - metadata
      properties:
        id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        list_price:
          type: string
        compare_at_price:
          nullable: true
          type: string
        currency:
          type: string
          maxLength: 3
        metadata:
          type: object
...
    CreateProductPriceRequest:
      type: object
      required:
        - website_id
        - item_id
        - list_price
        - currency
        - metadata
      properties:
        website_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        list_price:
          type: string
        compare_at_price:
          nullable: true
          type: string
        currency:
          type: string
          maxLength: 3
        metadata:
          type: object
...
    ProductPriceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductPrice'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    RecoveryInvite:
      type: object
      required:
        - id
        - cart_id
        - template_ref
        - delivery_state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        cart_id:
          type: string
          format: uuid
        template_ref:
          type: string
          maxLength: 200
        notified_at:
          nullable: true
          type: string
          format: date-time
        delivery_state:
          type: string
          maxLength: 16
        metadata:
          type: object
...
    CreateRecoveryInviteRequest:
      type: object
      required:
        - cart_id
        - template_ref
        - delivery_state
        - metadata
      properties:
        cart_id:
          type: string
          format: uuid
        template_ref:
          type: string
          maxLength: 200
        notified_at:
          nullable: true
          type: string
          format: date-time
        delivery_state:
          type: string
          maxLength: 16
        metadata:
          type: object
...
    RecoveryInviteList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecoveryInvite'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []